chore(WIP): major updates in typings and runtime
- Moved "data.numbers" to "dotnet.types"
- Moved "TypedNumber" to typing dotnet section
- Created "CompareExpression" to the reworked data.compare function
- Refactored "std.shift" function (only code internal changes)
- (WIP) Rework of "data.compare" in progress
