+++ projects/melon-runtime/Melon.Library/Standard/std.js
@@ -1,21 +1,18 @@
 ﻿const std = {
     Promise: Promise,
     shift: (value) => {
-        const internal = {
-            option: (target, callback) => {
-                switch (typeof target) {
-                    case "object":
-                        target.includes(value) && callback(target[target.indexOf(value)]);
-                        break;
+        return (target, callback) => {
+            switch (typeof target) {
+                case "object":
+                    target.includes(value) && callback(target[target.indexOf(value)]);
+                    break;
 
-                    default:
-                        value === target && callback(target);
-                        break;
-                }
-                return std.shift(value);
+                default:
+                    value === target && callback(target);
+                    break;
             }
+            return std.shift(value);
         }
-        return internal;
     },
     melon: {
         currentVersion: new Version(),
