+++ projects/melon-types/dotnet.d.ts
@@ -1,3 +1,4 @@
+import { TypedNumber } from "./types/dotnet/TypedNumber";
 import { DotnetFetchExpression } from "./types/dotnet/DotnetFetchExpression";
 import { DotnetGetStaticMethodResult } from "./types/dotnet/DotnetGetStaticMethodResult";
 import { RealmConstructor } from "./types/dotnet/RealmConstructor";
@@ -8,6 +9,19 @@ type Dotnet = {
     loadAssembly: (path: string) => string,
     removeAssembly: (path: string) => void,
     getAssemblies: (path: string) => (string | null)[],
+    types: {
+        SByte(value: number): TypedNumber,
+        Byte(value: number): TypedNumber,
+        Short(value: number): TypedNumber,
+        UShort(value: number): TypedNumber,
+        Int(value: number): TypedNumber,
+        UInt(value: number): TypedNumber,
+        Long(value: number): TypedNumber,
+        ULong(value: number): TypedNumber,
+        Float(value: number): TypedNumber,
+        Double(value: number): TypedNumber,
+        Decimal(value: number): TypedNumber
+    }
     Realm: RealmConstructor
 }
