+++ projects/melon-types/types/data/CompareExpression.ts
@@ -1,0 +1,3 @@
+type CompareExpression<T> = (t1: T, t2: T) => boolean;
+
+export { CompareExpression }
