--- projects/melon-types/types/data/CompareFunction.ts
@@ -1,3 +1,0 @@
-type CompareFunction<T> = (t1: T, t2: T) => boolean;
-
-export { CompareFunction }
