+++ projects/melon-types/types/data/CompareResult.ts
@@ -1,0 +1,7 @@
+type CompareResult = {
+    comment: string,
+    equals: boolean,
+    equalityPercent: number
+}
+
+export { CompareResult }
