+++ projects/melon-types/types/dotnet/TypedNumber.ts
